Added solution 29.
* Made the prime numbers iterator return 64-bit integers.
* Changed other solutions to match the same.
* Handled 0 in the GCD function.

+++ src/solutions/summation_of_primes.rs
@@ -2,7 +2,7 @@ use crate::utils;
 
 pub fn solve() {
     const LIMIT: usize = 2000000;
-    let result: usize = utils::primes(LIMIT).sum();
+    let result: i64 = utils::primes(LIMIT).sum();
 
     println!("{}", result);
     assert_eq!(result, 142913828922);
